requirements.txt
googletrans==4.0.0rc1
python-docx==0.8.11
python-magic==0.4.27
minio==7.1.17